refactor reports/templates
